[rename] directory name
